Создал компоненты пустой и заполненой страницы фаворитных офферов и реализовал логику
